pyproject.toml
[project]
name = "autenticapje"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "clear>=2.0.0",
    "cryptography>=46.0.3",
    "jpype1>=1.6.0",
    "pandas>=2.3.3",
    "pykeepass>=4.1.1.post1",
    "python-dotenv>=1.2.1",
    "selenium>=4.38.0",
    "tqdm>=4.67.1",
    "typer>=0.20.0",
    "webdriver-manager>=4.0.2",
    "werkzeug>=3.1.4",
]
